+++ test/test_ul4.py
@@ -423,10 +423,10 @@ all_templates = dict(
 	python=TemplatePython,
 	python_dumps=TemplatePythonDumpS,
 	python_dump=TemplatePythonDump,
-	# java_compiled_by_python=TemplateJavaCompiledByPython,
-	# java_compiled_by_java=TemplateJavaCompiledByJava,
-	# js_v8=TemplateJavascriptV8,
-	# js_spidermonkey=TemplateJavascriptSpidermonkey,
+	java_compiled_by_python=TemplateJavaCompiledByPython,
+	java_compiled_by_java=TemplateJavaCompiledByJava,
+	js_v8=TemplateJavascriptV8,
+	js_spidermonkey=TemplateJavascriptSpidermonkey,
 	# php=TemplatePHP,
 )